refactor: move 5 scoring modules into scoring/
