+++ benchmarks/bench_forward_recall.py
@@ -76,7 +76,7 @@ from helix_context.schemas import (
     Gene,
     PromoterTags,
 )
-from helix_context.tcm import SessionContext, tcm_bonus
+from helix_context.scoring.tcm import SessionContext, tcm_bonus
 
 
 SHARED_DOMAIN = "ephemeron"
