+++ tests/test_sr.py
@@ -18,7 +18,7 @@ class _FakeGenome:
 @pytest.fixture(autouse=True)
 def _patch_loader(monkeypatch):
     """Redirect _load_co_activated to the fake graph so we don't need a DB."""
-    from helix_context import ray_trace
+    from helix_context.scoring import ray_trace
 
     def fake_load(genome, gene_id):
         return genome._graph.get(gene_id, [])
